Updated incomplete url and added form click action
